Organized Files / Fixed POM
