+++ pom.xml
@@ -19,6 +19,17 @@
             </plugin>
         </plugins>
     </build>
+
+    <properties>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
+        <mysql-connector-version>8.0.30</mysql-connector-version>
+        <owner-version>1.0.12</owner-version>
+        <log4j-version>2.18.0</log4j-version>
+        <jaxb-version>2.3.0</jaxb-version>
+        <jackson-version>2.13.4</jackson-version>
+    </properties>
+
     <dependencies>
         <dependency>
             <groupId>org.apache.logging.log4j</groupId>
@@ -35,5 +46,11 @@
             <artifactId>mysql-connector-java</artifactId>
             <version>8.0.30</version>
         </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+            <version>2.10.0</version>
+        </dependency>
     </dependencies>
+
 </project>
