+++ src/main/java/src/main/java/dao/AttorneyDAO.java
@@ -1,6 +1,6 @@
 package src.main.java.dao;
 
-import src.main.java.interfaces.IAttorneyDAO;
+import src.main.java.utils.interfaces.IAttorneyDAO;
 import src.main.java.model.Attorney;
 
 import java.util.List;
