+++ src/main/java/src/main/java/dao/BillingDAO.java
@@ -1,6 +1,6 @@
 package src.main.java.dao;
 
-import src.main.java.interfaces.IBaseDAO;
+import src.main.java.utils.interfaces.IBaseDAO;
 
 public class BillingDAO implements IBaseDAO {
     @Override
