+++ src/main/java/src/main/java/dao/UserDao.java
@@ -1,6 +1,6 @@
 package src.main.java.dao;
 
-import src.main.java.interfaces.IUserDAO;
+import src.main.java.utils.interfaces.IUserDAO;
 import src.main.java.model.User;
 
 import java.util.ArrayList;
