--- src/main/java/src/main/java/service/Main.java
@@ -1,8 +1,0 @@
-package src.main.java.service;
-
-public class Main {
-
-    public static void main(String[] args) {
-
-    }
-}
