+++ src/main/java/src/main/java/services/Main.java
@@ -1,0 +1,15 @@
+package src.main.java.services;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+public class Main {
+    private static final Logger logger = LogManager.getLogger(Main.class);
+
+
+    public static void main(String[] args) {
+
+
+
+    }
+}
