+++ src/main/java/src/main/java/utils/interfaces/IConnectionPool.java
@@ -1,6 +1,6 @@
-package src.main.java.interfaces;
+package src.main.java.utils.interfaces;
 
-import src.main.java.service.ConnectionPool;
+import src.main.java.utils.ConnectionPool;
 
 import java.sql.Connection;
