+++ src/main/resources/json/BillingJson.java
@@ -1,22 +1,24 @@
-package src.main.java.json;
+package src.main.java.resources.json;
 
 import org.json.simple.JSONObject;
 
 import java.io.FileWriter;
 import java.io.IOException;
 
-public class CourthouseJson {
+public class BillingJson {
 
     public static void main(String args[]) {
 
         JSONObject jsonObject = new JSONObject();
 
-        jsonObject.put("category_id", "1");
-        jsonObject.put("Name", "Albuquerque Courthouse");
-        jsonObject.put("Attorney_id", "1");
+        jsonObject.put("id", "1");
+        jsonObject.put("Base_cost", "100");
+        jsonObject.put("Hourly_rate", "50");
+        jsonObject.put("Billingcol", " ");
+        jsonObject.put("Payment_id", "1");
 
         try {
-            FileWriter file = new FileWriter("E:/Courthouseoutput.json");
+            FileWriter file = new FileWriter("E:/Billingoutput.json");
             file.write(jsonObject.toJSONString());
             file.close();
         } catch (IOException e) {
