+++ src/main/resources/json/CourthouseJson.java
@@ -1,22 +1,22 @@
-package src.main.java.json;
+package src.main.java.resources.json;
 
 import org.json.simple.JSONObject;
 
 import java.io.FileWriter;
 import java.io.IOException;
 
-public class TestimonialsJson {
+public class CourthouseJson {
 
     public static void main(String args[]) {
 
         JSONObject jsonObject = new JSONObject();
 
-        jsonObject.put("id", "1");
-        jsonObject.put("rating", "10");
-
+        jsonObject.put("category_id", "1");
+        jsonObject.put("Name", "Albuquerque Courthouse");
+        jsonObject.put("Attorney_id", "1");
 
         try {
-            FileWriter file = new FileWriter("E:/Testimonialsoutput.json");
+            FileWriter file = new FileWriter("E:/Courthouseoutput.json");
             file.write(jsonObject.toJSONString());
             file.close();
         } catch (IOException e) {
