+++ src/main/resources/json/TestimonialsJson.java
@@ -1,24 +1,22 @@
-package src.main.java.json;
+package src.main.java.resources.json;
 
 import org.json.simple.JSONObject;
 
 import java.io.FileWriter;
 import java.io.IOException;
 
-public class BillingJson {
+public class TestimonialsJson {
 
     public static void main(String args[]) {
 
         JSONObject jsonObject = new JSONObject();
 
         jsonObject.put("id", "1");
-        jsonObject.put("Base_cost", "100");
-        jsonObject.put("Hourly_rate", "50");
-        jsonObject.put("Billingcol", " ");
-        jsonObject.put("Payment_id", "1");
+        jsonObject.put("rating", "10");
+
 
         try {
-            FileWriter file = new FileWriter("E:/Billingoutput.json");
+            FileWriter file = new FileWriter("E:/Testimonialsoutput.json");
             file.write(jsonObject.toJSONString());
             file.close();
         } catch (IOException e) {
