+++ src/main/resources/xsd/Consultation.xsd
